Moved Hotkey class to new file
