+++ civ5-keybinder/civ5-keybinder/MainWindow.xaml
@@ -24,7 +24,11 @@
                     <ColumnDefinition Width="*"/>
                 </Grid.ColumnDefinitions>
 
-                <Button Grid.Column="0" Margin="0 0 5 5" Name="applyButton" Content="Apply" Click="applyButton_Click"/>
+                <Button Grid.Column="0" Margin="0 0 5 5" Name="applyButton" Content="Apply" Click="applyButton_Click">
+                    <!--<Button.InputBindings>
+                        <KeyBinding Command="{Binding Path=button_PreviewKeyDown}" Key="A"/>
+                    </Button.InputBindings>-->
+                </Button>
                 <Button Grid.Column="1" Margin="5 0 0 5" Name="resetButton" Content="Reset to Defaults" Click="resetButton_Click"/>
 
             </Grid>
