+++ civ5-keybinder/civ5-keybinder/MainWindow.xaml.cs
@@ -26,10 +26,10 @@ namespace civ5_keybinder
 
             InitializeGrid(DefaultHotkeys());
         }
-        
+
         public void InitializeGrid(List<Hotkey> hotkeys)
         {
-            // I don't know why adding this additional row is neccesary, but it is
+            // I don't know why this additional row is neccesary, but it is
             main_grid.RowDefinitions.Add(new RowDefinition());
 
             void AddText(string text, int row, int col)
@@ -48,7 +48,7 @@ namespace civ5_keybinder
                 button.Content = text;
                 button.GotKeyboardFocus += new KeyboardFocusChangedEventHandler(button_GotKeyboardFocus);
                 button.LostKeyboardFocus += new KeyboardFocusChangedEventHandler(button_LostKeyboardFocus);
-                button.KeyDown += new KeyEventHandler(button_KeyDown);
+                button.PreviewKeyDown += new KeyEventHandler(button_PreviewKeyDown);
                 Grid.SetRow(button, row);
                 Grid.SetColumn(button, col);
                 main_grid.Children.Add(button);
@@ -92,15 +92,78 @@ namespace civ5_keybinder
                             AddCheckBox(hotkeys[hotkey_num].Alt, hotkey_num + 1, 5);
                             break;
                     }
-                }                
+                }
             }
         }
 
-        private void button_KeyDown(object sender, KeyEventArgs e)
+        private void button_PreviewKeyDown(object sender, KeyEventArgs e)
         {
             Button button = sender as Button;
-            object initialContent = button.Content;
-            button.Content = e.Key.ToString().ToUpper();
+
+            //MessageBox.Show(e.Key.ToString());
+
+            //Dictionary<Key, string> specialKeys = new Dictionary<Key, string>
+            //{
+            //    {Key.PageDown, "PageDown"},
+            //    {Key.PageUp, "PageUp"},
+            //    //{Key.System, "F10"},
+            //    {Key.D0, "0"},
+            //};
+
+            if (e.Key == Key.Return)
+            {
+                button.Content = "Enter";
+                e.Handled = true;
+            }
+            else if (e.Key == Key.System)
+            {
+                button.Content = "F10";
+                e.Handled = true;
+            }
+            else if (e.Key == Key.Next)
+            {
+                button.Content = "PageDown";
+                // Prevents page from scrolling when pressed
+                e.Handled = true;
+            }
+            else if (e.Key == Key.PageDown)
+            {
+                button.Content = "PageUp";
+                e.Handled = true;
+            }
+            //else if (e.Key == Key.Space)
+            //{
+            //    button.Content = "hello";
+            //}
+            else
+            {
+                if (e.Key.ToString().Length > 1)
+                {
+                    button.Content = e.Key.ToString();
+                    e.Handled = true;
+                }
+                else
+                {
+                    button.Content = e.Key.ToString().ToUpper();
+                    e.Handled = true;
+                }
+            }
+
+            //if (specialKeys.ContainsKey(e.Key) == true)
+            //{
+            //    button.Content = specialKeys[e.Key];
+            //}
+            //else
+            //{
+            //    if (e.Key.ToString().Length > 1)
+            //    {
+            //        button.Content = e.Key.ToString();
+            //    }
+            //    else
+            //    {
+            //        button.Content = e.Key.ToString().ToUpper();
+            //    }
+            //}
         }
 
         private void button_GotKeyboardFocus(object sender, KeyboardFocusChangedEventArgs e)
@@ -237,47 +300,6 @@ namespace civ5_keybinder
         private void resetButton_Click(object sender, RoutedEventArgs e)
         {
             MessageBox.Show("reset button clicked");
-        }
-    }
-
-    public class Hotkey
-    {
-        public int File { get; }
-        public string DLC { get; }
-        public string Function { get; }
-        public string Key { get; set; }
-        public bool Ctrl { get; set; }
-        public bool Shift { get; set; }
-        public bool Alt { get; set; }
-
-        public Hotkey(int defFile, int defDLC, string defFunction, string defKey, bool defCtrl, bool defShift, bool defAlt)
-        {
-            File = defFile;
-            switch (defDLC)
-            {
-                case 0:
-                    DLC = "";
-                    break;
-                case 1:
-                    DLC = "G + K";
-                    break;
-                case 2:
-                    DLC = "BNW";
-                    break;
-            }
-            Function = defFunction;
-            Key = defKey;
-            Ctrl = defCtrl;
-            Shift = defShift;
-            Alt = defAlt;
-        }
-
-        public void DefineBinding(string defKey, bool defCtrl, bool defShift, bool defAlt)
-        {
-            Key = defKey;
-            Ctrl = defCtrl;
-            Shift = defShift;
-            Alt = defAlt;
         }
     }
 }
